requirements.txt
pandas
numpy
pyarrow
faiss-cpu
sentence-transformers
python-dotenv
openai
streamlit
tqdm
accelerate
transformers 
streamlit
fastapi
uvicorn


